Move initializeState out of legacy
