+++ packages/satcheljs/lib/index.ts
@@ -2,20 +2,20 @@ export { default as ActionCreator } from './interfaces/ActionCreator';
 export { default as ActionMessage } from './interfaces/ActionMessage';
 export { default as Subscriber } from './interfaces/Subscriber';
 export { actionCreator, boundActionCreator } from './actionCreator';
+export { default as createStore } from './createStore';
 export { dispatch } from './dispatcher';
+export { default as initializeState } from './initializeState';
 export { mutator, registerMutators } from './mutator';
 export { orchestrator, registerOrchestrators } from './orchestrator';
 export { default as rootStore } from './rootStore';
 export { useStrict } from 'mobx';
 
 // Legacy API
-export { default as initializeState } from './legacy/initializeState';
 export { default as applyMiddleware } from './legacy/applyMiddleware';
 export { default as Middleware } from './legacy/Middleware';
 export { default as ActionFunction } from './legacy/ActionFunction';
 export { default as ActionContext } from './legacy/ActionContext';
 export { default as DispatchFunction } from './legacy/DispatchFunction';
-export { default as createStore } from './createStore';
 export { default as action } from './legacy/action';
 export { default as select, SelectorFunction } from './legacy/select';
 export { default as createUndo, UndoResult, CreateUndoReturnValue} from './legacy/createUndo';
