+++ packages/satcheljs/test/legacy/dispatchTests.ts
@@ -2,7 +2,7 @@ import 'jasmine';
 import { action as mobxAction, autorun, _ } from 'mobx';
 
 import rootStore from '../../lib/rootStore';
-import initializeState from '../../lib/legacy/initializeState';
+import initializeState from '../../lib/initializeState';
 import dispatch from '../../lib/legacy/dispatch';
 import * as applyMiddlewareImports from '../../lib/legacy/applyMiddleware';
 import { __resetGlobalContext } from '../../lib/globalContext'
