+++ packages/satcheljs/test/legacy/initializeStateTests.ts
@@ -1,6 +1,6 @@
 import 'jasmine';
 import rootStore from '../../lib/rootStore';
-import initializeState from '../../lib/legacy/initializeState';
+import initializeState from '../../lib/initializeState';
 
 describe("initializeState", () => {
     it("replaces the state value", () => {
